UI, volunteers db, update profile

+++ AndroidStudioProjects/AidHarbour/app/src/main/java/com/example/aidharbour/ItemsViewModel.kt
@@ -1,4 +1,4 @@
 package com.example.aidharbour
 
-class ItemsViewModel (val name: String, val rating: Int){
+class ItemsViewModel (val name: String, val rating: Double){
 }
